jpa/jpashop/src/test/java/wtchrs/SpringMockShop/repository/MemberRepositoryTest.java
package wtchrs.SpringMockShop.repository;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.transaction.annotation.Transactional;
import wtchrs.SpringMockShop.domain.Member;

import static org.assertj.core.api.Assertions.*;

@SpringBootTest
class MemberRepositoryTest {

    @Autowired
    private MemberRepository memberRepository;

    @Test
    @Transactional
    @Rollback(false)
    void saveAndFind() {
        Member member = new Member("hello", null);
        Long saveId = memberRepository.save(member);

        Member findMember = memberRepository.findById(saveId);

        assertThat(findMember.getId()).isEqualTo(member.getId());
        assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
    }
}